Panic when loading or parsing of pages or posts fails during build
This will prevent silently ignoring accidental errors. If a user has many posts/pages
and accidently breaks the header, previously mublog would fail parsing, and then silently skip
over the broken post/page, causing it not to show up in the final build.
Now, we immediatly error out, if we fail to parse or load a page/post.
Since we have posts/pages in dedicated directories now, this is fine, since we don't expect
there to be other markdown files in those directories (such as READMEs) that could cause a build failure.