Major refactor of Context architecture
The BlogContext struct was kind of a helper construct, that just contains everything.
It was used in the pipeline, stored all the posts, pages, assets, path configs etc.
This lead to a very poor design an messy code.
This commit is the first step towards cleaning up the BlogContext structure, and having it
only contain a ComponentRegistry and a PathConfig. These containers can then be used to
access the things we need instead.