Apply PostListingFeature config options
This commit enables sorting the post listings by date in either
ascending or descending order. The sorting is determined by the option
set in the Mublog configuration file. Configuration options for
enabled features are automatically parsed when parsing the config file.
Each feature has a dedicated config struct with feature-specific
options. The current implementation requires copying the feature
configuration in the feature stage because the context, which
contains it, is passed as a mutable reference.
This is not ideal, and should be improved in future iterations.