Pipeline error propagation + Navbar ordering
This commit ensures that errors, that occur in the execution of a
pipeline stage or feature are propagated all the way the stack.
That makes understanding errors that occur during the build process
much easier.

Furthermore support for ordering links generated by the NavbarFeature
based on the mublog config file is now possible.

Note: This commit introduced a lot of ugly hacks. Major refactoring
will be necessary.